requirements.txt
Flask==3.1.3
gunicorn==23.0.0
python-docx==1.2.0
lxml==5.3.1
typing_extensions==4.12.2
